+++ wiwb/api_calls/__init__.py
@@ -2,5 +2,6 @@ from wiwb.api_calls.base import Request
 from wiwb.api_calls.get_data_sources import GetDataSources
 from wiwb.api_calls.get_grids import GetGrids
 from wiwb.api_calls.get_variables import GetVariables
+from wiwb.api_calls.body import RequestBody, ReaderSettings, Interval, Extent, Exporter, Reader, ExporterSettings
 
-__all__ = ["Request", "GetDataSources", "GetVariables", "GetGrids"]
+__all__ = ["Request", "GetDataSources", "GetVariables", "GetGrids", "RequestBody", "ReaderSettings", "Interval", "Extent", "Exporter", "Reader", "ExporterSettings"]
