+++ wiwb/api_calls/get_grids.py
@@ -12,6 +12,7 @@ from pandas import DataFrame
 from shapely.geometry import MultiPolygon, Point, Polygon
 
 from wiwb.api_calls import Request
+from wiwb.api_calls.body import RequestBody, ReaderSettings, Interval, Extent, Exporter, Reader
 from wiwb.constants import (
     DATA_FORMAT_CODES,
     FILE_SUFFICES,
@@ -23,261 +24,6 @@ from wiwb.sample import sample_netcdf
 
 logger = logging.getLogger(__name__)
 defaults = get_defaults()
-
-
-@dataclass
-class Extent:
-    f"""Extent for Settings in request body in correct epsg: {defaults.crs}.
-
-    Parameters
-    ----------
-    xll : float
-        The x-coordinate of the lower-left corner of the extent. Defaults to {defaults.bounds[0]}.
-    yll : float
-        The y-coordinate of the lower-left corner of the extent. Defaults to {defaults.bounds[1]}.
-    xur : float
-        The x-coordinate of the upper-right corner of the extent. Defaults to {defaults.bounds[2]}.
-    yur : float
-        The y-coordinate of the upper-right corner of the extent. Defaults to {defaults.bounds[3]}.
-    """
-
-    xll: float = defaults.bounds[0]
-    yll: float = defaults.bounds[1]
-    xur: float = defaults.bounds[2]
-    yur: float = defaults.bounds[3]
-
-    def __post_init__(self):
-        if self.width <= 0:
-            raise ValueError(
-                f"'xll' ({self.xll}) should be smaller than 'xur' ({self.xur})"
-            )
-
-        if self.height <= 0:
-            raise ValueError(
-                f"'yll' ({self.yll}) should be smaller than 'yur' ({self.yur})"
-            )
-
-        self.correct_bounds()
-
-    @property
-    def width(self):
-        return self.xur - self.xll
-
-    @property
-    def height(self):
-        return self.yur - self.yll
-
-    @property
-    def crs(self):
-        return pyproj.CRS(self.epsg)
-
-    @property
-    def epsg(self):
-        return defaults.crs
-
-    @property
-    def spatial_reference(self):
-        return {"Epsg": self.epsg}
-
-    def correct_bounds(self):
-        # get crs-unit
-        units = None
-        crs_dict = self.crs.to_dict()
-        if "unit" in crs_dict.keys():
-            units = crs_dict["units"]
-
-        # get min width and height
-        if units == "m":
-            min_width_height = 10
-        else:
-            min_width_height = 0.0001  # we assume degrees
-
-        # alter bounds
-        if self.width < min_width_height:
-            logger.warning(
-                f"""Width of bounds < min_width ({self.width < min_width_height}). {self.xll} and {self.xur} will be adjusted"""  # noqa:E501
-            )
-            self.xll -= (min_width_height - self.width) / 2
-            self.xur += (min_width_height - self.width) / 2
-
-        if self.height < min_width_height:
-            logger.warning(
-                f"""Height of bounds < min_height ({self.height < min_width_height}). {self.yll} and {self.yur} will be adjusted"""  # noqa:E501
-            )
-            self.yll -= (min_width_height - self.height) / 2
-            self.yur += (min_width_height - self.height) / 2
-
-    def json(self):
-        dict = self.__dict__.copy()
-        dict["spatial_reference"] = self.spatial_reference
-        return {snake_to_pascal_case(k): v for k, v in dict.items()}
-
-
-@dataclass
-class Interval:
-    """Interval for Settings in request body.
-
-    Parameters
-    ----------
-    type : str
-        The interval, either "Days", "Hours", "Minutes"
-    value: int
-        Increment of the interval
-
-    Example
-    -------
-
-    Interval(type="Hours", value=2)
-
-    Is an interval of 2 hours.
-
-    """
-
-    type: INTERVAL_TYPES
-    value: int
-
-    def __post_init__(self):
-        self.validate()
-
-    def validate(self):
-        if self.type not in INTERVAL_TYPES.__args__:
-            raise ValueError(
-                f"{self.type} not a valid interval-type: {INTERVAL_TYPES.__args__}"
-            )
-
-    def json(self):
-        self.validate()
-        return {snake_to_pascal_case(k): v for k, v in self.__dict__.items()}
-
-
-@dataclass
-class ReaderSettings:
-    """WIWB reader-settings
-
-    Parameters
-    ----------
-    start_date: datetime.date
-        Reader start_date
-    end_date: datetime.date
-        Reader end_date
-    model_date: datetime.date
-        Reader end_date
-    variable_codes: List[str]
-        List of WIWB variable codes
-    interval: Interval
-        time-interval for reader
-    extent: Extend
-        extent for reader
-    """
-
-    start_date: date
-    end_date: date
-    variable_codes: list
-    model_date: Union[date | None] = None
-    interval: Union[Interval, None] = None
-    extent: Union[Extent, None] = field(default_factory=Extent)
-
-    def json(self):
-        dict = self.__dict__.copy()
-        dict["start_date"] = dict["start_date"].strftime("%Y%m%d%H%M%S")
-        dict["end_date"] = dict["end_date"].strftime("%Y%m%d%H%M%S")
-        for k in ["interval", "extent", "model_date"]:
-            if dict[k] is None:
-                dict.pop(k)
-            elif k == "model_date":
-                dict[k] = dict["model_date"].strftime("%Y%m%d%H%M%S")
-            else:
-                dict[k] = dict[k].json()
-
-        return {snake_to_pascal_case(k): v for k, v in dict.items() if v is not None}
-
-
-@dataclass
-class Reader:
-    """WIWB reader
-
-    Parameters
-    ----------
-    data_source_code: str
-        WIWB datasourcecode to read
-    settings: Settings
-        WIWB reader settings
-    """
-
-    data_source_code: str
-    settings: Union[ReaderSettings, None] = field(default_factory=ReaderSettings)
-
-    def json(self):
-        dict = self.__dict__.copy()
-        dict["settings"] = dict["settings"].json()
-
-        return {snake_to_pascal_case(k): v for k, v in dict.items()}
-
-
-@dataclass
-class ExporterSettings:
-    """WIWB export settings
-
-    Parameters
-    ----------
-    export_projection_file: bool, optional
-        To write a projection file (in case of ASCII Grid). Default is False
-
-    """
-
-    export_projection_file = False
-
-
-@dataclass
-class Exporter:
-    f"""WIWB exporter
-
-    Parameters
-    ----------
-    data_format_code: {DATA_FORMAT_CODES}, optional
-        data-format code to export data to. Defaults to geotiff
-    settings: ExporterSettings
-        WIWB exporter settings
-
-    """
-
-    data_format_code: DATA_FORMAT_CODES = "geotiff"
-    settings: Union[ExporterSettings, None] = None
-
-    def __post_init__(self):
-        self.validate()
-
-    def validate(self):
-        if self.data_format_code not in DATA_FORMAT_CODES.__args__:
-            raise ValueError(
-                f"{self.data_format_code} not a valid data-format-code: {DATA_FORMAT_CODES.__args__}"
-            )
-
-    def json(self):
-        self.validate()
-        return {
-            snake_to_pascal_case(k): v
-            for k, v in self.__dict__.items()
-            if v is not None
-        }
-
-
-@dataclass
-class RequestBody:
-    """GetGrids request Body"""
-
-    readers: List[Reader]
-    exporter: Union[Exporter, None] = None
-
-    def json(self):
-        dict = self.__dict__.copy()
-        dict["readers"] = [i.json() for i in dict["readers"]]
-        if dict["exporter"] is None:
-            dict.pop("exporter")
-        else:
-            dict["exporter"] = dict["exporter"].json()
-        return {snake_to_pascal_case(k): v for k, v in dict.items()}
-
 
 @dataclass
 class GetGrids(Request):
